Allow specifying the seed for RNG
As a result, the Parameters class no longer holds pointers to RNG
objects, only the seed. The actual RNG is constructed in ITPSystem.
This makes the design clearer, and provides a way to re-run identical
simulations.

This commit also includes a command line parser flag for specifying the
RNG seed. The seed is also saved in datafiles for easy reproduction
later.